requirements.txt
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
pandasgui==0.2.14
prophet==1.1.6
scikit-learn==1.6.1
statsmodels==0.14.4
openai~=1.70.0
python-dotenv~=1.1.0